backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
cohere==4.34
qdrant-client>=1.7.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
pdfplumber==0.9.0
pytest==7.4.3
asyncpg>=0.29.0